Return early from various search functions when the target is empty
I noticed that if I went into a search/replace dialog Ox would freeze if
I tried to find all instances of an empty string.  This PR checks to see
if the string given in both the search prompt, the replace prompt, and
next_match/prev_match is empty.  If so, return from the current function
early.